+++ src/main/scala/sangria/macros/package.scala
@@ -16,7 +16,9 @@ package object macros {
   }
 
 
-  def deriveObjectType[Ctx, Val](config: DeriveConfig[Ctx, Val]*): ObjectType[Ctx, Val] = macro DeriveMacro.deriveObjectType[Ctx, Val]
+  def deriveObjectType[Ctx, Val](config: DeriveObjectTypeConfig[Ctx, Val]*): ObjectType[Ctx, Val] = macro DeriveMacro.deriveObjectType[Ctx, Val]
+
+  def deriveEnum[T](config: DeriveEnumTypeConfig*): Int = macro DeriveMacro.deriveEnumType[T]
 
   class GraphQLName(name: String) extends StaticAnnotation
   class GraphQLDescription(description: String) extends StaticAnnotation